Added methods get/setAdditionalData for MCRFile and MCRDirectory objects, removed unneccessary casts etc.
git-svn-id: https://server.mycore.de/svn/mycore/trunk@5868 7089d063-bc14-496d-ac90-53de1872eb6c

+++ sources/org/mycore/datamodel/ifs/MCRAudioVideoExtender.java
@@ -128,12 +128,10 @@ public abstract class MCRAudioVideoExtender {
      */
     public String getBitRateFormatted() {
         if (bitRate > (1024 * 1024)) {
-            double b = (double) (Math.round((double) bitRate / 10485.76)) / 100.0;
-
+            double b = Math.round(bitRate / 10485.76) / 100.0;
             return new DecimalFormat("##0.##").format(b) + " MBit";
         } else {
-            double b = (double) (Math.round((double) bitRate / 102.4)) / 10.0;
-
+            double b = Math.round(bitRate / 102.4) / 10.0;
             return new DecimalFormat("##0.#").format(b) + " kBit";
         }
     }
